Fix forecast-horizon feature generation: climate, alert features, Rt/p_rt1
- Climate: forecast-horizon tempmed/humidmed now use city + ISO-week-of-year
  historical mean instead of flat carry-forward of the last known value.
- Drop transmissao/receptivo/nivel_inc as model input features (cases
  models and the epidemic classifier) -- these are InfoDengue-internal
  alert-classifier outputs with no documented formula, so they can't be
  computed for future weeks. nivel_inc is kept only as the classifier's
  label and as a side-channel for the legacy benchmark comparison.
- Replace the CI-regime proxy with the trained epidemic classifier's
  predicted probability instead of raw nivel_inc, since the classifier can
  run forward on forecasted weeks. Added train_final_classifier/
  select_best_classifier, wired classifier nested CV into the main
  pipeline, and joined its OOF probability into the existing calibration
  machinery (conditional_residuals.attach_classifier_proxy).
- Add features/rt_estimation.py: a from-scratch numpy/scipy port of the
  Codeco et al. (2017) temperature-dependent Wallinga-Teunis Rt estimator
  (no R/rpy2 dependency). Validated against the paper's own published
  example (FIdata.csv) and against the real R implementation run on actual
  InfoDengue data (R0 package installed via conda for this comparison).
  Wired into forecast_next_52w.py via a two-pass design: a draft
  cases-forecast pass feeds a one-shot Rt/p_rt1 estimate over the full
  horizon, then a final pass uses that estimate instead of carry-forward.
- Fix two latent bugs surfaced along the way: sustained_rt's bool dtype
  upcasting to `object` after .shift() (rejected by XGBoost), and
  negative-valued draft case predictions breaking the renewal-equation's
  non-negativity requirement (now clipped before Rt estimation).
- Add ml/src/dengue_ml/features/Codeco-et-al-2017/ reference materials
  (paper PDF, original R scripts, example datasets) for future reference.

+++ ml/src/dengue_ml/preprocessing.py
@@ -91,10 +91,15 @@ def prepare_weekly_table(
     # a real outbreak (75.5% recall, 57% precision against true nivel_inc==2
     # onsets) than the pre-epidemic alert level alone (100% recall, 11.7%
     # precision) -- worth exposing to the model as its own feature.
+    # float (0.0/1.0), not bool: a bool column's .shift() in
+    # add_weekly_lag_features introduces NaN for the first few rows per city,
+    # which pandas can only represent by upcasting the whole column to
+    # `object` dtype -- XGBoost then rejects it ("DataFrame.dtypes ... must
+    # be int, float, bool or category").
     high_rt = df["p_rt1"] > 0.95
     df["sustained_rt"] = (
         high_rt.groupby(df[CITY_COL]).transform(lambda s: s.rolling(3).sum() >= 3)
-    )
+    ).astype(float)
 
     df = df.rename(columns={"umidmed": "humidmed"})
     cols = [
